hubble/container: Properly deal with nil values in RingReader
When the ring-buffer is not full, we must treat nil values the same as
reads outside of the readable indices. This commit ensures that `read`
treats nil values the same way `readFrom` does.

This has an impact on the RingReader, as we should not try to either
increase or decrease the index if the current value is unreadable.
If the current index is unreadable, then we have reached either end
of the ring buffer.

This also adds an error return type to `RingReader.{Next,Prev}` in
order to determine which end we have reached.

Most notably, `Next` may return ErrConcurrentRead. This means this
means that the value it tried to access just got overwritten by the
writer from behind. In such cases we want to return an error, as
otherwise `RingReader.Next` would silently stop reading and
subsequentially terminate the request as if it had reached most
recent value in the ring buffer (which it has not).

The error is also added to `Previous`, however in that case
we only ever expect ErrConcurrentRead, as it traverses the ring buffer
in the opposite direction of the writer.

The check for concurrent writes is currently missing from `readFrom`,
which can suffer from the same problem. This will be added in a
follow-up PR.

Fixes: cilium/hubble#131

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>